Refactor into new library servant-conctest
