+++ servant-conctest.cabal
@@ -18,25 +18,32 @@ source-repository head
   type: git
   location: https://github.com/plow-technologies/servant-conctest.git
 
+library
+  exposed-modules:
+    Servant.ConcTest
+  other-modules:
+    Servant.ConcTest.Utils
+  hs-source-dirs: src
+  build-depends:
+      base
+    , async
+    , http-client
+    , QuickCheck
+    , servant-server
+    , servant-client
+    , warp
+  default-language: Haskell2010
+
 executable example
   main-is: Main.hs
-  hs-source-dirs:
-      app
-  -- ghc-options: -Wall -Wincomplete-uni-patterns -Wincomplete-record-updates -Wunused-packages -O2 -threaded -with-rtsopts=-N
+  hs-source-dirs: app
+  -- TODO try with O2?
   ghc-options: -Wall -Wincomplete-uni-patterns -Wincomplete-record-updates -Wunused-packages -threaded -with-rtsopts=-N
   build-depends:
       base
-    -- , containers
-    , async
-    -- , hspec
     , http-client
-    -- , pqc -- NOTE use my fork
     , QuickCheck
-    -- , servant-quickcheck
+    , servant-conctest
     , servant-server
     , servant-client
-    -- , string-conversions
-    -- , text
-    -- , wai-logger
-    , warp
   default-language: Haskell2010
